add `9vx`/`vx32`
`9vx` is a binary emulator for linux allowing users to run plan9 binaries on linux.
It also allows for extracting the 9front iso into a folder and running 9front from a folder on the linux file system.

Since `mame` is in the repo I thought it would be a good idea to add `9vx` too since it allows for some really cool grid networking.